pyproject.toml
[project]
name = "controvert"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Daniel Rodriguez", email = "darc2203@gmail.com" }
]
dependencies = [
    "neo4j>=5.26.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
    "tqdm>=4.67.0",
    "yfinance>=0.2.50",
    "streamlit>=1.40.1",
    "plotly>=5.24.1",
    "tabulate>=0.9.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/controvert"]
